requirements.txt
Flask==2.0.1
python-decouple==3.4
Flask-RESTful==0.3.9
uwsgi